refactor cors internals
