+++ docs/2.utils/98.advanced.md
@@ -159,38 +159,6 @@ Make a proxy request to a target URL and send the response back to the client.
 
 <!-- automd:jsdocs src="../../src/utils/cors.ts" -->
 
-### `appendCorsHeaders(event, options)`
-
-Append CORS headers to the response.
-
-### `appendCorsPreflightHeaders(event, options)`
-
-Append CORS preflight headers to the response.
-
-### `createAllowHeaderHeaders(event, options)`
-
-Create the `access-control-allow-headers` and `vary` headers.
-
-### `createCredentialsHeaders(options)`
-
-Create the `access-control-allow-credentials` header.
-
-### `createExposeHeaders(options)`
-
-Create the `access-control-expose-headers` header.
-
-### `createMaxAgeHeader(options)`
-
-Create the `access-control-max-age` header.
-
-### `createMethodsHeaders(options)`
-
-Create the `access-control-allow-methods` header.
-
-### `createOriginHeaders(event, options)`
-
-Create the `access-control-allow-origin` header.
-
 ### `handleCors(event, options)`
 
 Handle CORS for the incoming request.
@@ -220,18 +188,6 @@ router.use('/',
  })
 );
 ```
-
-### `isCorsOriginAllowed(origin, options)`
-
-Check if the incoming request is a CORS request.
-
-### `isPreflightRequest(event)`
-
-Check if the incoming request is a CORS preflight request.
-
-### `resolveCorsOptions(options)`
-
-Resolve CORS options.
 
 <!-- /automd -->
